✅ Complete location_filters.py modularization
- Created map_locations/location_filters.py with 8 functions
- Reduced core.py from 2,275 to 2,067 lines (-208 lines)
- Updated __init__.py to import from location_filters module
- All backward compatibility maintained
- Testing: ✅ All imports and functionality work correctly

Progress: 50% Complete (3/6 modules)
🎯 ALL PRIORITY 1 FOUNDATION MODULES COMPLETE! 🎯
✅ categories.py ✅ utils.py ✅ location_filters.py
Next: popup_generators.py (Priority 2: Medium Risk)
