+++ map_locations/core_modularization_plan.md
@@ -48,6 +48,25 @@ The `core.py` file has grown to 2,552 lines and handles multiple responsibilitie
 - [x] `create_sample_locations()`
 
 
+### 📂 location_filters.py - COMPLETED ✅
+- **Completed**: December 2024
+- **Lines Moved**: 218 lines removed from core.py
+- **Functions Migrated**: 8 functions
+- **Testing**: ✅ All imports work correctly
+- **Backward Compatibility**: ✅ Maintained via __init__.py
+- **Core.py Size**: Reduced from 2,275 to 2,067 lines (-208 lines net)
+
+**Functions Successfully Migrated:**
+- [x] `get_available_types()`
+- [x] `get_available_tags()`
+- [x] `get_available_neighborhoods()`
+- [x] `filter_locations_by_type()`
+- [x] `filter_locations_by_tags()`
+- [x] `filter_locations_by_neighborhood()`
+- [x] `get_location_summary()`
+- [x] `validate_location_data()`
+
+
 ## Module Breakdown
 
 ### ✅ Phase 1: Create Module Structure
@@ -80,6 +99,25 @@ The `core.py` file has grown to 2,552 lines and handles multiple responsibilitie
 
 **Functions Successfully Migrated:**
 - [x] `create_sample_locations()`
+
+
+### 📂 location_filters.py - COMPLETED ✅
+- **Completed**: December 2024
+- **Lines Moved**: 218 lines removed from core.py
+- **Functions Migrated**: 8 functions
+- **Testing**: ✅ All imports work correctly
+- **Backward Compatibility**: ✅ Maintained via __init__.py
+- **Core.py Size**: Reduced from 2,275 to 2,067 lines (-208 lines net)
+
+**Functions Successfully Migrated:**
+- [x] `get_available_types()`
+- [x] `get_available_tags()`
+- [x] `get_available_neighborhoods()`
+- [x] `filter_locations_by_type()`
+- [x] `filter_locations_by_tags()`
+- [x] `filter_locations_by_neighborhood()`
+- [x] `get_location_summary()`
+- [x] `validate_location_data()`
 
 **Imports needed:**
 ```python
@@ -168,19 +206,19 @@ from .common import LocationList
 ---
 
 #### 5. 📂 `location_filters.py` - Location Analysis & Filtering
-- **Lines**: ~185-339, 2170-2388 (scattered, ~400 lines total)
-- **Status**: ⏳ Pending
+- **Lines**: ~1979-2195 (218 lines)
+- **Status**: ✅ COMPLETED
 - **Dependencies**: `categories.py`
 
 **Functions to move:**
-- [ ] `get_available_types()`
-- [ ] `get_available_tags()`
-- [ ] `get_available_neighborhoods()`
-- [ ] `filter_locations_by_type()`
-- [ ] `filter_locations_by_tags()`
-- [ ] `filter_locations_by_neighborhood()`
-- [ ] `get_location_summary()`
-- [ ] `validate_location_data()`
+- [x] `get_available_types()`
+- [x] `get_available_tags()`
+- [x] `get_available_neighborhoods()`
+- [x] `filter_locations_by_type()`
+- [x] `filter_locations_by_tags()`
+- [x] `filter_locations_by_neighborhood()`
+- [x] `get_location_summary()`
+- [x] `validate_location_data()`
 
 **Imports needed:**
 ```python
@@ -199,6 +237,25 @@ from .common import LocationList
 
 **Functions to move:**
 - [x] `create_sample_locations()`
+
+
+### 📂 location_filters.py - COMPLETED ✅
+- **Completed**: December 2024
+- **Lines Moved**: 218 lines removed from core.py
+- **Functions Migrated**: 8 functions
+- **Testing**: ✅ All imports work correctly
+- **Backward Compatibility**: ✅ Maintained via __init__.py
+- **Core.py Size**: Reduced from 2,275 to 2,067 lines (-208 lines net)
+
+**Functions Successfully Migrated:**
+- [x] `get_available_types()`
+- [x] `get_available_tags()`
+- [x] `get_available_neighborhoods()`
+- [x] `filter_locations_by_type()`
+- [x] `filter_locations_by_tags()`
+- [x] `filter_locations_by_neighborhood()`
+- [x] `get_location_summary()`
+- [x] `validate_location_data()`
 
 **Imports needed:**
 ```python
@@ -376,11 +433,11 @@ cp core_backup.py core.py
 
 ## Progress Tracking
 
-### Current Progress: 33% Complete (2/6 modules)
+### Current Progress: 50% Complete (3/6 modules)
 
 - [x] Analysis completed
 - [x] Plan documented
-- [x] Foundation modules created (2/3: categories.py ✅, utils.py ✅)
+- [x] Foundation modules created (3/3: categories.py ✅, utils.py ✅, location_filters.py ✅)
 - [ ] Content modules created
 - [ ] Complex modules created
 - [ ] Import structure updated
